+++ modules/compiler/include/compiler/module.h
@@ -350,8 +350,109 @@ struct KernelInfo {
   }
 };  // class KernelInfo
 
-/// @brief Kernel info callback type.
-using KernelInfoCallback = std::function<void(KernelInfo)>;
+/// @brief Class for managing program information.
+///
+/// Owns instances of KernelInfo (e.g., one per kernel in a module/program) and
+/// provides utilities for iterating over them.
+struct ProgramInfo {
+  /// @brief Add a single kernel info.
+  bool addNewKernel(KernelInfo &&info) {
+    if (cargo::success != kernel_descriptions.emplace_back(std::move(info))) {
+      return false;
+    }
+    return true;
+  }
+
+  /// @brief Initialize empty program information for a specified number of
+  /// kernels for later population.
+  ///
+  /// @param[in] numKernels Number of kernels to allocate space for.
+  bool resizeFromNumKernels(int32_t numKernels) {
+    if (cargo::success != kernel_descriptions.resize(numKernels)) {
+      return false;
+    }
+    return true;
+  }
+
+  inline size_t getNumKernels() const { return kernel_descriptions.size(); }
+
+  /// @brief Retrieve a kernel by index.
+  ///
+  /// @param[in] kernel_index Index into the list of kernel infos.
+  ///
+  /// @return Return kernel info if found, null otherwise.
+  KernelInfo *getKernel(size_t kernel_index) {
+    if (kernel_index >= kernel_descriptions.size()) {
+      return nullptr;
+    }
+    return &kernel_descriptions[kernel_index];
+  }
+
+  /// @brief Retrieve a kernel by index.
+  ///
+  /// @param[in] kernel_index Index into the list of kernel infos.
+  ///
+  /// @return Return kernel info if found, null otherwise.
+  const KernelInfo *getKernel(size_t kernel_index) const {
+    if (kernel_index >= kernel_descriptions.size()) {
+      return nullptr;
+    }
+    return &kernel_descriptions[kernel_index];
+  }
+
+  /// @brief Retrieve a kernel by name.
+  ///
+  /// @param[in] kernel_name Name of the kernel to search for.
+  ///
+  /// @return Return kernel info if found, null otherwise.
+  compiler::KernelInfo *getKernelByName(cargo::string_view kernel_name) {
+    for (auto &desc : kernel_descriptions) {
+      if (kernel_name == desc.name) {
+        return &desc;
+      }
+    }
+    return nullptr;
+  }
+
+  /// @brief Retrieve a kernel by name.
+  ///
+  /// @param[in] kernel_name Name of the kernel to search for.
+  ///
+  /// @return Return kernel info if found, null otherwise.
+  const compiler::KernelInfo *getKernelByName(
+      cargo::string_view kernel_name) const {
+    for (const auto &desc : kernel_descriptions) {
+      if (kernel_name == desc.name) {
+        return &desc;
+      }
+    }
+    return nullptr;
+  }
+
+  /// @brief Retrieve the begin iterator.
+  ///
+  /// @return The beginning of the kernel info range.
+  KernelInfo *begin() { return kernel_descriptions.begin(); }
+
+  /// @brief Retrieve the begin iterator.
+  ///
+  /// @return The beginning of the kernel info range.
+  const KernelInfo *begin() const { return kernel_descriptions.begin(); }
+
+  /// @brief Retrieve the end iterator.
+  ///
+  /// @return Return the end iterator.
+  KernelInfo *end() { return kernel_descriptions.end(); }
+
+  /// @brief Retrieve the end iterator.
+  ///
+  /// @return Return the end iterator.
+  const KernelInfo *end() const { return kernel_descriptions.end(); }
+
+ private:
+  /// @brief Kernel descriptions.
+  cargo::small_vector<KernelInfo, 8> kernel_descriptions;
+};
 
 namespace spirv {
 /// @brief Information about the target device to used during SPIR-V
@@ -513,7 +614,7 @@ class Module {
 
   /// @brief Generates a binary from the current program.
   ///
-  /// @param[out] kernel_info_callback Kernel info callback.
+  /// @param[out] program_info Optional ProgramInfo object to fill in.
   /// @param[out] printf_calls Output printf descriptor list.
   ///
   /// @return Return a status code.
@@ -522,7 +623,7 @@ class Module {
   /// @retval `Result::FINALIZE_PROGRAM_FAILURE` when finalization failed. See
   /// the error log for more information.
   virtual Result finalize(
-      KernelInfoCallback kernel_info_callback,
+      ProgramInfo *program_info,
       std::vector<builtins::printf::descriptor> &printf_calls) = 0;
 
   /// @brief Creates a binary from the current module. This assumes that the